+++ finite-volume/source/cuda/common.h
@@ -27,16 +27,16 @@
 */
 
 // available device functions
-void cuda_smooth(level_type level, int x_id, int rhs_id, double a, double b, int s, double *c, double *d);
-void cuda_residual(level_type d_level, int res_id, int x_id, int rhs_id, double a, double b);
-void cuda_rebuild(level_type level, int x_id, int Aii_id, int sumAbsAij_id, double a, double b);
+void cuda_smooth(level_type level, int x_id, int rhs_id, float a, float b, int s, float *c, float *d);
+void cuda_residual(level_type d_level, int res_id, int x_id, int rhs_id, float a, float b);
+void cuda_rebuild(level_type level, int x_id, int Aii_id, int sumAbsAij_id, float a, float b);
 
 void cuda_restriction(level_type d_level_c, int id_c, level_type d_level_f, int id_f, communicator_type restriction, int restrictionType, int block_type);
 
-void cuda_interpolation_p0(level_type d_level_f, int id_f, double prescale_f, level_type d_level_c, int id_c, communicator_type interpolation, int block_type);
-void cuda_interpolation_p1(level_type d_level_f, int id_f, double prescale_f, level_type d_level_c, int id_c, communicator_type interpolation, int block_type);
-void cuda_interpolation_v2(level_type level_f, int id_f, double prescale_f, level_type level_c, int id_c, communicator_type interpolation, int block_type);
-void cuda_interpolation_v4(level_type level_f, int id_f, double prescale_f, level_type level_c, int id_c, communicator_type interpolation, int block_type);
+void cuda_interpolation_p0(level_type d_level_f, int id_f, float prescale_f, level_type d_level_c, int id_c, communicator_type interpolation, int block_type);
+void cuda_interpolation_p1(level_type d_level_f, int id_f, float prescale_f, level_type d_level_c, int id_c, communicator_type interpolation, int block_type);
+void cuda_interpolation_v2(level_type level_f, int id_f, float prescale_f, level_type level_c, int id_c, communicator_type interpolation, int block_type);
+void cuda_interpolation_v4(level_type level_f, int id_f, float prescale_f, level_type level_c, int id_c, communicator_type interpolation, int block_type);
 
 void cuda_apply_BCs_v1(level_type level, int x_id, int shape);
 void cuda_apply_BCs_v2(level_type level, int x_id, int shape);
@@ -44,15 +44,15 @@ void cuda_apply_BCs_v4(level_type level, int x_id, int shape);
 void cuda_extrapolate_betas(level_type level, int shape);
 
 void cuda_zero_vector(level_type d_level, int id);
-void cuda_scale_vector(level_type d_level, int id_c, double scale_a, int id_a);
-void cuda_shift_vector(level_type d_level, int id_c, double shift_a, int id_a);
-void cuda_mul_vectors(level_type d_level, int id_c, double scale, int id_a, int id_b);
-void cuda_add_vectors(level_type d_level, int id_c, double scale_a, int id_a, double scale_b, int id_b);
-double cuda_sum(level_type d_level, int id);
-double cuda_max_abs(level_type d_level, int id);
+void cuda_scale_vector(level_type d_level, int id_c, float scale_a, int id_a);
+void cuda_shift_vector(level_type d_level, int id_c, float shift_a, int id_a);
+void cuda_mul_vectors(level_type d_level, int id_c, float scale, int id_a, int id_b);
+void cuda_add_vectors(level_type d_level, int id_c, float scale_a, int id_a, float scale_b, int id_b);
+float cuda_sum(level_type d_level, int id);
+float cuda_max_abs(level_type d_level, int id);
 void cuda_color_vector(level_type d_level, int id_a, int colors_in_each_dim, int icolor, int jcolor, int kcolor);
 
 void cuda_copy_block(level_type d_level, int id, communicator_type exchange_ghosts, int block_type);
-void cuda_increment_block(level_type d_level, int id, double prescale, communicator_type exchange_ghosts, int block_type);
+void cuda_increment_block(level_type d_level, int id, float prescale, communicator_type exchange_ghosts, int block_type);
 
 #include "extra.h"
