+++ finite-volume/source/timers/mpi.c
@@ -5,6 +5,6 @@
 //------------------------------------------------------------------------------------------------------------------------------
 #include <stdint.h>
 #include <mpi.h>
-double getTime(){
+float getTime(){
   return(MPI_Wtime()); // timers are in units of seconds; no conversion is necessary
 }
