+++ finite-volume/source/timers/omp.c
@@ -5,6 +5,6 @@
 //------------------------------------------------------------------------------------------------------------------------------
 #include <stdint.h>
 #include <omp.h>
-double getTime(){
+float getTime(){
   return(omp_get_wtime()); // timers are in units of seconds; no conversion is necessary
 }
